+++ hss_okta_flutter/ios/Classes/HssOktaFlutterPlugin.swift
@@ -15,8 +15,6 @@ case generalError(String)
 
 
 public class HssOktaFlutterPlugin: NSObject, FlutterPlugin,HssOktaFlutterPluginApi {
-  
-    
 
     let browserAuth = WebAuthentication.shared
     var flow : (any AuthenticationFlow)?
@@ -63,6 +61,9 @@ public class HssOktaFlutterPlugin: NSObject, FlutterPlugin,HssOktaFlutterPluginA
                             
                         case .mfaRequired(_):
                             completion(.success(AuthenticationResult(result: DirectAuthenticationResult.mfaRequired)))
+                        default:
+                            break
+                            
                         }
                     }else{
                         completion(.failure(HssOktaError.generalError("Incorrect flow")))
@@ -230,6 +231,8 @@ public class HssOktaFlutterPlugin: NSObject, FlutterPlugin,HssOktaFlutterPluginA
                                 resultEnum: nil,
                                 token: result.token, userInfo: userInfo)))
                         
+                        
+                        
                     }
                     
                     completion(.failure(HssOktaError.credentialError("Failed to save credentials")))
@@ -241,8 +244,56 @@ public class HssOktaFlutterPlugin: NSObject, FlutterPlugin,HssOktaFlutterPluginA
         }
     }
     
+    func getAllUserIds(completion: @escaping (Result<[String], Error>) -> Void) {
+        Task{
+            let ids = Credential.allIDs
+            completion(.success(ids))
+        }
+    }
     
+    func getToken(tokenId: String, completion: @escaping (Result<AuthenticationResult?, Error>) -> Void) {
+        Task{
+            do{
+                if let fetchedCredential = try Credential.with(id: tokenId){
+                    completion(.success(constructAuthenticationResult(
+                        resultEnum: nil, token: fetchedCredential.token, userInfo: fetchedCredential.userInfo)))
+                }
+               
+            }catch let error{
+                completion(.failure(error))
+            }
+        }
+    }
     
+   
+    
+    func removeCredential(tokenId: String, completion: @escaping (Result<Bool, Error>) -> Void) {
+        Task{
+            Task{
+                do{
+                    
+                    try Credential.with(id:tokenId)?.remove()
+                    completion(.success(true))
+                    
+                }catch let error{
+                    completion(.failure(error))
+                }
+            }
+        }
+    }
+    
+    func setDefaultToken(tokenId: String, completion: @escaping (Result<Bool, Error>) -> Void) {
+        Task{
+            do{
+               
+                try Credential.tokenStorage.setDefaultTokenID(tokenId)
+                completion(.success(true))
+                
+            }catch let error{
+                completion(.failure(error))
+            }
+        }}
+
     
     //    Helper methods