+++ hss_okta_flutter/CHANGELOG.md
@@ -64,3 +64,18 @@
 * Added User claims class
 * Added Session API
 * exposed AuthClient for web users
+
+# 1.3.3
+* Added Fix for a problem in iOS which causes non-exhaustive error, making some users unable to build
+
+# 1.3.4
+* Added getAllIds for iOS and Android
+* Added removeCredential for iOS and Android
+* Added setDefaultCredential for iOS
+* Added getToken for Android and iOS
+* Fixed Missing error return for native Browser SignIn and Signout streams
+* Fixed Crash on android when okta.properties has bad values
+
+# 1.4.0
+* Fixed renew token not accepting abstract tokens
+* updated example UI